requirements.txt
Flask==0.9
Flask-SQLAlchemy==0.16
Jinja2==2.7
Mako==0.8.0
MarkupSafe==0.18
SQLAlchemy==0.8.1
Werkzeug==0.8.3
alembic==0.5.0
amqplib==1.0.2
argparse==1.2.1
distribute==0.6.24
mongoengine==0.8.1
pymongo==2.5.2
pyserial==2.6
python-dateutil==2.1
pytz==2013b
redis==2.7.5
six==1.3.0
wsgiref==0.1.2
yolk==0.4.3
